Add files via upload
In the new version, I have abandoned many unnecessary widgets and their connections in favor of working with MainWindow and the status bar. Now there is the ability to search for a file. All actions with the file are reflected in the status bar. The file can be edited and sent for processing. There is still a problem with the license, but the "tmp.fc" file generated after processing can be analyzed when opened directly in the computational software complex.

---

If you have any more texts to translate or further questions, feel free to ask!